Move reason command's logic to suspensions command
